Moving non-util files out of utils package
