+++ core/src/test/scala/org/apache/spark/eventhubs/ConnectionStringBuilderSuite.scala
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.spark.eventhubs.utils
+package org.apache.spark.eventhubs
 
 import java.time.Duration
