+++ core/src/test/scala/org/apache/spark/eventhubs/EventHubsConfSuite.scala
@@ -17,11 +17,7 @@
 
 package org.apache.spark.eventhubs
 
-import org.apache.spark.eventhubs.utils.{
-  ConnectionStringBuilder,
-  EventHubsTestUtils,
-  EventPosition
-}
+import org.apache.spark.eventhubs.utils.EventHubsTestUtils
 import org.json4s.NoTypeHints
 import org.json4s.jackson.Serialization
 import org.json4s.jackson.Serialization.{ read => sread }
