+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/EventHubsDirectDStreamSuite.scala
@@ -23,10 +23,10 @@ import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.concurrent.atomic.{ AtomicInteger, AtomicLong }
 
 import com.microsoft.azure.eventhubs.EventData
-import org.apache.spark.eventhubs.EventHubsConf
+import org.apache.spark.eventhubs.{ EventHubsConf, EventPosition }
 import org.apache.spark.eventhubs.utils.EventHubsTestUtils._
 import org.apache.spark.eventhubs.rdd.{ HasOffsetRanges, OffsetRange }
-import org.apache.spark.eventhubs.utils.{ EventHubsTestUtils, EventPosition, SimulatedClient }
+import org.apache.spark.eventhubs.utils.{ EventHubsTestUtils, SimulatedClient }
 import org.apache.spark.{ SparkConf, SparkFunSuite }
 import org.apache.spark.internal.Logging
 import org.apache.spark.rdd.RDD
